# [Optimus] add einsum_to_pointwise_pass pattern